requirements.txt
dbt-core==1.9.4
dbt-snowflake==1.9.0
sqlfluff-templater-dbt==3.3.1
python-dotenv==1.0.1
scikit-learn==1.7.2
pandas==2.3.3
